Better handle allowedVehicleIndices in the two-step planner.
The original version used the intersection of allowed vehicle indices
for all shipments served from the parking location. When there are
significant variations in the allowed vehicles among the shipments, this
may lead to situations where the set of vehicles that can serve the
parking location is empty.

The modified version creates a new parking location visit for each group
of allowed vehicle indices used with the parking location. The new
version is equivalent to the original one when allowed vehicle indices
are not used, but it can create more shipments in the global model when
many different sets of allowed vehicles are used.