Add detours to the merged response.
For shipments picked up/delivered directly, the detours are taken from the
global model and they have the usual meaning.
For shipments delivered from a parking location, the detour is computed as the
sum of the detour in the global model (the extra time spent globally on the way
to the parking) + the detour in the local model (the extra time spent locally
in the parking).
For the virtual visits, the detours are defined as:
- for the arrival, it is the time difference compared to driving straight to
  the parking (the detour of the parking visit in the global model).
- for the departure, it is the time spent at the parking (the duration of the
  local route).