Added a refinement phase to the two-step routing library.
In this refinement phase, the solver collects all sequences of visits to
a parking location, and resolves them using a modified local model to
reoptimize for exactly the shipments delivered in these visits. It
replaces the base local routes with the refined ones and then re-solves
the global model by using the original solution as a warm-start.

This refinement has the following effects:
- it can merge visits from multiple delivery rounds in the original
  (unrefined) solution, thus removing unnecessary returns to the parking
  location to pick up more shipments.
- by removing these returns to parking, it can save some time and open
  up opportunities to deliver even more shipments in the saved time, or
  better organize shipments now that the delivery rounds are known.

We observed that it is very effective at reducing the inefficient
returns, but it can also help with reducing other types of parking
issues and even reduce skipped shipments.

After the two refinement phases, there might still be unnecessary visits
to the parking location - typically they would be created by the refined
global phase that can now perform some shipments that were skipped in
the original model.