Pin start time of delivery rounds in the local refinement requests.
We've seen cases where the solver managed to save a lot of time at the
end by starting a few seconds before the soft start time, which in turn
caused infeasibility of the first solution injected to the following
global refinement phase. Pinning the start time should avoid these
cases, even though it might prevent the solver from saving time by
pushing the start of the local delivery round to a later time.