Add visit start times to the injected first solution in global refinement steps.
This should help bootstrap the solver in global refinement requests, and will
be needed to terminate the refinement with a local request.

For now, we only added the code, but it is protected by a flag that is
turned off by default, i.e. there is no change in behavior.