[evaluate_solution] Added a script that computes the costs of a CFR JSON request+response.
The script takes a CFR JSON request and a (partial) response, and uses the CFR
API to compute the cost of a solution where the routes use exactly the visit
sequences from the input response. This is done by creating a CFR request where
the routes are constrained to the visits from the input using
injectedSolutionConstraint and where all the skipped shipments are removed.

Typical use:
  python3 -m cfr.json.evaluate_solution \
    --project "${YOUR_GCLOUD_PROJECT_ID}" \
    --token "${YOUR_GCLOUD_TOKEN}" \
    --input_request scenario.json \
    --input_response alternative_solution.json \
    --output_response alternative_solution_with_costs.json