[two_step_routing] Compute load and unload time at the parking location.
The load and unload times are specified per parking location, through
the `unload_duration_per_item` and `load_duration_per_item` attributes
of the ParkingLocation class. The `unload` attribute is for taking items
delivered from the parking out of the vehicle on arrival, the `load`
attribute is for putting items picked up through the parking location
into the vehicle before departure.

In the merged models, the loading and unloading time is exported as the
duration of the arrival-to-parking and departure-from-parking shipments.