[cfr_json_analysis] Added code to detect warping in GMPRO routes.
Warping is detected using route polylines in transitions between visit. The warp
distance is defined as the distance between the last point of the inbound polyline
and the first point of the outbound polyline, i.e. the distance between the arrival
coordinates and the departure coordinates.

Adds the necessary code to compute the distances to `analysis.py`, and adds the
statistics to vehicle-oriented cells in the analysis colab notebook.